+++ docs/CHANGELOG.md
@@ -64,6 +64,7 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Add `r_shadowquality [0-5]`, `r_shadowdistance [0-5]`, `r_shadowitems`, `r_shadowcorpses`, and `dbg_shadows` console commands
 - Add `r_vertexlighting` console command and `$Use Vertex Lighting` `MAPNAME_info.tbl` option to restore legacy vertex lighting for meshes
 - Alias `TAB` in level editor to toggle maximized viewport (stock hotkeys `F4`/`F5`)
+- In TDM match mode, made self kills reduce team score by 1 (floor of 0)
 
 [@is-this-c](https://github.com/is-this-c)
 - Use 64-bit integers for time deltas